+++ lib/job_enqueue_logger/delayed_job_adapter.rb
@@ -34,8 +34,8 @@ module JobEnqueueLogger
       alias server? server
 
       def logger
-        if !server? && defined?(Rails)
-          Rails.logger
+        if !server? && JobEnqueueLogger.logger
+          JobEnqueueLogger.logger
         else
           Delayed::Worker.logger
         end