+++ lib/job_enqueue_logger/railtie.rb
@@ -6,6 +6,7 @@ module JobEnqueueLogger
   class Railtie < Rails::Railtie
     initializer "job_enqueue_logger.backtrace_cleaner" do
       JobEnqueueLogger.backtrace_cleaner = Rails.backtrace_cleaner
+      JobEnqueueLogger.logger = Rails.logger
     end
   end
 end