+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## master (unreleased)
 
+- Add ability to set logger
 - Prefix this gem log lines with `[JobEnqueueLogger]`
 
 ## 0.2.0 (2023-04-04)